requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-cors-headers==2.2.0
django-extensions==2.0.7
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
oauthlib==2.1.0
openapi-codec==1.3.2
Pillow==5.1.0
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.15.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
sqlparse